% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1X.R
\docType{data}
\name{data1X}
\alias{data1X}
\title{A synthesized input variable dataset to fit a linear model on a panel dataset.}
\format{
A data frame with 5,000 rows and 2 columns:
\describe{
\item{columns}{the two independent variables}
\item{rows}{each 100 rows represent the timeseries of each of the 50 individuals}
...
}
}
\usage{
data1X
}
\description{
A synthesized input variable dataset to fit a linear model on a panel dataset.
}
\keyword{datasets}
