% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data1Y.R
\docType{data}
\name{data1Y}
\alias{data1Y}
\title{A synthesized output variable dataset to fit a linear model on a panel dataset.}
\format{
A data frame with 100 rows and 50 columns:
\describe{
\item{columns}{the individuals}
\item{rows}{the time points in the timeseries of each individual}
...
}
}
\usage{
data1Y
}
\description{
A synthesized output variable dataset to fit a linear model on a panel dataset.
}
\keyword{datasets}
