% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2X.R
\docType{data}
\name{data2X}
\alias{data2X}
\title{A synthesized input variable dataset to fit a binomial model on a panel dataset.}
\format{
A data frame with 5,000 rows and 2 columns:
\describe{
\item{columns}{the two independent variables}
\item{rows}{each 50 rows represent the timeseries of each of the 100 individuals}
...
}
}
\usage{
data2X
}
\description{
A synthesized input variable dataset to fit a binomial model on a panel dataset.
}
\keyword{datasets}
