% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3X.R
\docType{data}
\name{data3X}
\alias{data3X}
\title{A synthesized input variable dataset to fit a poisson model on a panel dataset.}
\format{
A data frame with 20,000 rows and 3 columns:
\describe{
\item{columns}{the three independent variables}
\item{rows}{each 100 rows represent the timeseries of each of the 200 individuals}
...
}
}
\usage{
data3X
}
\description{
A synthesized input variable dataset to fit a poisson model on a panel dataset.
}
\keyword{datasets}
