% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data3Y.R
\docType{data}
\name{data3Y}
\alias{data3Y}
\title{A synthesized output variable dataset to fit a poisson model on a panel dataset.}
\format{
A data frame with 100 rows and 200 columns:
\describe{
\item{columns}{the individuals}
\item{rows}{the time points in the timeseries of each individual}
...
}
}
\usage{
data3Y
}
\description{
A synthesized output variable dataset to fit a poisson model on a panel dataset.
}
\keyword{datasets}
