% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4LAB.R
\docType{data}
\name{data4LAB}
\alias{data4LAB}
\title{A synthesized vector of memberships needed to fit a linear model on a panel dataset under known group memberships.}
\format{
A vector with 300 entries indicating the group membership of each individual.
}
\usage{
data4LAB
}
\description{
A synthesized vector of memberships needed to fit a linear model on a panel dataset under known group memberships.
}
\keyword{datasets}
