% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4Y.R
\docType{data}
\name{data4Y}
\alias{data4Y}
\title{A synthesized output variable dataset to fit a linear model on a panel dataset under known group memberships.}
\format{
A data frame with 100 rows and 300 columns:
\describe{
\item{columns}{the individuals}
\item{rows}{the time points in the timeseries of each individual}
...
}
}
\usage{
data4Y
}
\description{
A synthesized output variable dataset to fit a linear model on a panel dataset under known group memberships.
}
\keyword{datasets}
