% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data5Y.R
\docType{data}
\name{data5Y}
\alias{data5Y}
\title{A synthesized output variable dataset to cluster individuals by heterogeneous panel data models with interactive effects.}
\format{
A data frame with 100 rows and 300 columns:
\describe{
\item{columns}{the individuals}
\item{rows}{the time points in the timeseries of each individual}
...
}
}
\usage{
data5Y
}
\description{
A synthesized output variable dataset to cluster individuals by heterogeneous panel data models with interactive effects.
}
\keyword{datasets}
