% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data7X.R
\docType{data}
\name{data7X}
\alias{data7X}
\title{A synthesized input variable dataset to fit a quantile panel data model on a panel dataset.}
\format{
A data frame with 20,000 rows and 8 columns:
\describe{
\item{columns}{the eight independent variables}
\item{rows}{each 100 rows represent the timeseries of each of the 200 individuals}
...
}
}
\usage{
data7X
}
\description{
A synthesized input variable dataset to fit a quantile panel data model on a panel dataset.
}
\keyword{datasets}
