% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data8Y.R
\docType{data}
\name{data8Y}
\alias{data8Y}
\title{A synthesized output variable dataset to fit a quantile VAR model with interactive effects and lag=2.}
\format{
A data frame with 302 rows and 30 columns:
\describe{
\item{columns}{the individuals}
\item{rows}{the time points in the timeseries of each individual}
...
}
}
\usage{
data8Y
}
\description{
A synthesized output variable dataset to fit a quantile VAR model with interactive effects and lag=2.
}
\keyword{datasets}
