% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HYPTEST.R
\name{HYPTEST}
\alias{HYPTEST}
\title{HYPTEST}
\usage{
HYPTEST(
  B,
  B0,
  Se,
  test = "two",
  individuals = seq(1, ncol(B)),
  variables = seq(1, nrow(B))
)
}
\arguments{
\item{B}{A dataframe of Coefficients as obtained in the output of any function in the package.}

\item{B0}{A dataframe of hypothetical coefficients to be evaluated in the test. (nrows should match number of variables and ncols should match number of individuals)}

\item{Se}{A dataframe of Standard Errors as obtained in the output of any function in the package.}

\item{test}{A string to determine what kind of test to run ("two" for two-tailed, "right" for right-tailed and "left for left-tailed).}

\item{individuals}{A list of individuals whose coefficients are to be tested. Default is all individuals in the B dataframe.}

\item{variables}{A list of variables whose coefficients are to be tested. Default is all variables in the B dataframe.}
}
\value{
A dataframe of p-values resulting from each individual test.
}
\description{
This function undergoes hypothesis testing for regression coefficients obtained from the various functions in the package.
}
\examples{
N <- 400
P <- 300

R <- 3
p <- 3

LAM <- matrix(rnorm(P*R,0,1),nrow=P,ncol=R)
FAC <- matrix(rnorm(N*R,0,1),nrow=N,ncol=R)
AY <- FAC\%*\%t(LAM)

AX <- matrix(rnorm(p*P*N,0,1),nrow=P*N)
for(j in 1:P){
  AX[(N*(j-1)+1):(N*j),1] <- AX[(N*(j-1)+1):(N*j),1]+0.15*AY[,j]
  AX[(N*(j-1)+1):(N*j),3] <- AX[(N*(j-1)+1):(N*j),3]-0.21*AY[,j]
}

B <- c(-1,2,-1)
AB <- B\%*\%t(rep(1,len=P))+matrix(0.1*runif(p*P,-1,1),p,P)
for(j in 1:P){
  AY[,j] <- AY[,j]+(AX[(N*(j-1)+1):(N*j),])\%*\%AB[,j]
}

PROB <- exp(AY)/(1+exp(AY))
AY <- trunc(matrix(runif(N*P,0,1),ncol=P)+PROB)
fit <- PDMIFLOGIT(AX,AY,R)
HYPTEST(fit$Coefficients,data.frame(c(0,1),c(-1,2)),fit$Se,"two",c(1,2),c(1,3))
}
