% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDMIFCLUST.R
\name{PDMIFCLUST}
\alias{PDMIFCLUST}
\title{PDMIFCLUST}
\usage{
PDMIFCLUST(X, Y, NGfactors, NLfactors, Maxit = 100, tol = 0.001)
}
\arguments{
\item{X}{The (NT) times p design matrix, without an intercept where N=number of individuals, T=length of time series, p=number of explanatory variables.}

\item{Y}{The T times N panel of response where N=number of individuals, T=length of time series.}

\item{NGfactors}{A pre-specified number of common factors across groups (see example).}

\item{NLfactors}{A pre-specified number of factors in each groups (see example).}

\item{Maxit}{A maximum number of iterations in optimization. Default is 100.}

\item{tol}{Tolerance level of convergence. Default is 0.001.}
}
\value{
A list with the following components:
\itemize{
\item Label: The estimated group membership for each of the individuals.
\item Coefficients: The estimated heterogeneous coefficients.
\item Lower05: Lower end (5\%) of the 90\% confidence interval of the regression coefficients.
\item Upper95: Upper end (95\%) of the 90\% confidence interval of the regression coefficients.
\item GlobalFactors: The estimated common factors across groups.
\item GlobalLoadings: The estimated factor loadings for the common factors.
\item GroupFactors: The estimated group-specific factors.
\item GroupLoadings: The estimated factor loadings for each group.
\item pval: p-value for testing hypothesis on heterogeneous coefficients.
\item Se: Standard error of the estimated regression coefficients.
}
}
\description{
Under a pre-specified number of groups and the number of common factors, this function implements clustering for N individuals in the panels.
Each of individuals in the group are subject to the group-specific unobserved common factors.
}
\examples{
N <- 200
NGroup <- 3
Rs <- rep(2,len=NGroup)
PP <- rep(200,len=NGroup)
P <- sum(PP)
p <- 3
R <- 2

AY <- matrix(0,nrow=N,ncol=P)
LAM <- matrix(rnorm(P*R,0,1),nrow=P,ncol=R)
FAC <- matrix(rnorm(N*R,0,1),nrow=N,ncol=R)
XG <- FAC\%*\%t(LAM)
XL <- matrix(0,ncol=P,nrow=N)

for(i in 1:NGroup){
  LAM <- matrix(rnorm(PP[i]*Rs[i],1,1),nrow=PP[i],ncol=Rs[i])
  FAC <- matrix(rnorm(N*Rs[i],0,1),nrow=N,ncol=Rs[i])
  X <- FAC\%*\%t(LAM)
  XL[,(PP[i]*(i-1)+1):(PP[i]*i)] <- X
}

ERR <- matrix(rnorm(N*P,0,1),nrow=N,ncol=P)
AY <- XG+XL+ERR
AX <- matrix(runif(p*P*N,-2,2),nrow=P*N)
AB <- matrix(rnorm(p*P,4,2),ncol=P)
for(j in 1:P){AY[,j] <- AY[,j]+AX[(N*(j-1)+1):(N*j),]\%*\%AB[,j]}

PDMIFCLUST(AX,AY,R,Rs)
}
\references{
Ando, T. and Bai, J. (2016) Panel data models with grouped factor structure under unknown group membership Journal of Applied Econometrics, 31, 163-191.

Ando, T. and Bai, J. (2017) Clustering huge number of financial time series: A panel data approach with high-dimensional predictors and factor structures. Journal of the American Statistical Association, 112, 1182-1198.
}
