% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDMIFCOUNT.R
\name{PDMIFCOUNT}
\alias{PDMIFCOUNT}
\title{PDMIFCOUNT}
\usage{
PDMIFCOUNT(X, Y, Nfactors, Maxit = 100, tol = 0.001)
}
\arguments{
\item{X}{The (NT) times p design matrix, without an intercept where N=number of individuals, T=length of time series, p=number of explanatory variables.}

\item{Y}{The T times N panel of response where N=number of individuals, T=length of time series.}

\item{Nfactors}{A pre-specified number of common factors.}

\item{Maxit}{A maximum number of iterations in optimization. Default is 100.}

\item{tol}{Tolerance level of convergence. Default is 0.001.}
}
\value{
A list with the following components:
\itemize{
\item Coefficients: The estimated heterogeneous coefficients.
\item Lower05: Lower end (5\%) of the 90\% confidence interval of the regression coefficients.
\item Upper95: Upper end (95\%) of the 90\% confidence interval of the regression coefficients.
\item Factors: The estimated common factors across groups.
\item Loadings: The estimated factor loadings for the common factors.
\item Predict: The conditional expectation of response variable.
\item pval: p-value for testing hypothesis on heterogeneous coefficients.
\item Se: Standard error of the estimated regression coefficients.
}
}
\description{
Under a known group membership, this function estimates heterogeneous poisson panel data models with interactive effects.
}
\examples{
N <- 400
P <- 300
R <- 3
p <- 3

LAM <- matrix(runif(P*R,0,1),nrow=P,ncol=R)
FAC <- matrix(runif(N*R,0,1),nrow=N,ncol=R)
AY <- FAC\%*\%t(LAM)
AX <- matrix(runif(p*P*N,0,1),nrow=P*N)

for(j in 1:P){
  AX[(N*(j-1)+1):(N*j),1] <- AX[(N*(j-1)+1):(N*j),1]+0.01*AY[,j]
  AX[(N*(j-1)+1):(N*j),3] <- AX[(N*(j-1)+1):(N*j),3]-0.01*AY[,j]
}

B <- c(-1,2,-1)
AB <- B\%*\%t(rep(1,len=P))

for(j in 1:P){
  AY[,j] <- AY[,j]+(AX[(N*(j-1)+1):(N*j),])\%*\%AB[,j]
}

MM <- exp(AY)
for(i in 1:N){
for(j in 1:P){
AY[i,j] <- rpois(1,lambda=MM[i,j])
}}

fit <- PDMIFCOUNT(AX,AY,R)
}
\references{
Ando, T., Bai, J. and Li, K. (2021) Bayesian and maximum likelihood analysis of large-scale panel choice models with unobserved heterogeneity, Journal of Econometrics.
}
