% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PDMIFLING.R
\name{PDMIFLING}
\alias{PDMIFLING}
\title{PDMIFLING}
\usage{
PDMIFLING(X, Y, Membership, NGfactors, NLfactors, Maxit = 100, tol = 0.001)
}
\arguments{
\item{X}{The (NT) times p design matrix, without an intercept where N=number of individuals, T=length of time series, p=number of explanatory variables.}

\item{Y}{The T times N panel of response where N=number of individuals, T=length of time series.}

\item{Membership}{A pre-specified group membership.}

\item{NGfactors}{A pre-specified number of common factors across groups (see example).}

\item{NLfactors}{A pre-specified number of factors in each groups (see example).}

\item{Maxit}{A maximum number of iterations in optimization. Default is 100.}

\item{tol}{Tolerance level of convergence. Default is 0.001.}
}
\value{
A list with the following components:
\itemize{
\item Coefficients: The estimated heterogeneous coefficients.
\item Lower05: Lower end (5\%) of the 90\% confidence interval of the regression coefficients.
\item Upper95: Upper end (95\%) of the 90\% confidence interval of the regression coefficients.
\item GlobalFactors: The estimated common factors across groups.
\item GlobalLoadings: The estimated factor loadings for the common factors.
\item GroupFactors: The estimated group-specific factors.
\item GroupLoadings: The estimated factor loadings for each group.
\item pval: p-value for testing hypothesis on heterogeneous coefficients.
\item Se: Standard error of the estimated regression coefficients.
}
}
\description{
Under a known group membership, this function estimates heterogeneous panel data models with interactive effects.
Together with the regression coefficients, this function estimates the unobserved common factor structures both for across/within groups.
}
\examples{
N <- 200
NGroup <- 3
Rs <- rep(2,len=NGroup)
PP <- rep(200,len=NGroup)
P <- sum(PP)
p <- 3
R <- 2

AY <- matrix(0,nrow=N,ncol=P)
LAB <- sort(rep(1:NGroup,len=P))
LAM <- matrix(rnorm(P*R,0,1),nrow=P,ncol=R)
FAC <- matrix(rnorm(N*R,0,1),nrow=N,ncol=R)
XG <- FAC\%*\%t(LAM)
XL <- matrix(0,ncol=P,nrow=N)

for(i in 1:NGroup){
  LAM <- matrix(rnorm(PP[i]*Rs[i],1,1),nrow=PP[i],ncol=Rs[i])
  FAC <- matrix(rnorm(N*Rs[i],0,1),nrow=N,ncol=Rs[i])
  X <- FAC\%*\%t(LAM)
  XL[,(PP[i]*(i-1)+1):(PP[i]*i)] <- X
}

ERR <- matrix(rnorm(N*P,0,1),nrow=N,ncol=P)
AY <- XG+XL+ERR
AX <- matrix(runif(p*P*N,-2,2),nrow=P*N)
AB <- matrix(rnorm(p*P,4,2),ncol=P)
for(j in 1:P){AY[,j] <- AY[,j]+AX[(N*(j-1)+1):(N*j),]\%*\%AB[,j]}

PDMIFLING(AX,AY,LAB,R,Rs)
}
\references{
Ando, T. and Bai, J. (2015) Asset Pricing with a General Multifactor Structure Journal of Financial Econometrics, 13, 556-604.
}
