% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2X.R
\docType{data}
\name{data2X}
\alias{data2X}
\title{A synthesized input variable dataset to fit a binomial model on a panel dataset.}
\format{
A data frame with 60,000 rows and 3 columns:
\describe{
\item{columns}{the three independent variables}
\item{rows}{each 300 rows represent the timeseries of each of the 200 individuals}
...
}
}
\usage{
data2X
}
\description{
A synthesized input variable dataset to fit a binomial model on a panel dataset.
}
\keyword{datasets}
