% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2Y.R
\docType{data}
\name{data2Y}
\alias{data2Y}
\title{A synthesized output variable dataset to fit a binomial model on a panel dataset.}
\format{
A data frame with 300 rows and 200 columns:
\describe{
\item{columns}{the individuals}
\item{rows}{the time points in the timeseries of each individual}
...
}
}
\usage{
data2Y
}
\description{
A synthesized output variable dataset to fit a binomial model on a panel dataset.
}
\keyword{datasets}
