% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data4X.R
\docType{data}
\name{data4X}
\alias{data4X}
\title{A synthesized input variable dataset to fit a linear model on a panel dataset under known group memberships.}
\format{
A data frame with 30,000 rows and 2 columns:
\describe{
\item{columns}{the two independent variables}
\item{rows}{each 100 rows represent the timeseries of each of the 300 individuals}
...
}
}
\usage{
data4X
}
\description{
A synthesized input variable dataset to fit a linear model on a panel dataset under known group memberships.
}
\keyword{datasets}
