% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data5X.R
\docType{data}
\name{data5X}
\alias{data5X}
\title{A synthesized input variable dataset to cluster individuals by heterogeneous panel data models with interactive effects.}
\format{
A data frame with 120,000 rows and 3 columns:
\describe{
\item{columns}{the three independent variables}
\item{rows}{each 200 rows represent the timeseries of each of the 600 individuals}
...
}
}
\usage{
data5X
}
\description{
A synthesized input variable dataset to cluster individuals by heterogeneous panel data models with interactive effects.
}
\keyword{datasets}
