% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data6X.R
\docType{data}
\name{data6X}
\alias{data6X}
\title{A synthesized input variable dataset to cluster individual units by nonlinear heterogeneous panel data models with interactive effects when the group membership is unknown}
\format{
A data frame with 960,000 rows and 4 columns:
\describe{
\item{columns}{the four independent variables}
\item{rows}{each 800 rows represent the timeseries of each of the 1200 individuals}
...
}
}
\usage{
data6X
}
\description{
A synthesized input variable dataset to cluster individual units by nonlinear heterogeneous panel data models with interactive effects when the group membership is unknown
}
\keyword{datasets}
