% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data6Y.R
\docType{data}
\name{data6Y}
\alias{data6Y}
\title{A synthesized output variable dataset to cluster individual units by nonlinear heterogeneous panel data models with interactive effects when the group membership is unknown.}
\format{
A data frame with 50 rows and 90 columns:
\describe{
\item{columns}{the individuals}
\item{rows}{the time points in the timeseries of each individual}
...
}
}
\usage{
data6Y
}
\description{
A synthesized output variable dataset to cluster individual units by nonlinear heterogeneous panel data models with interactive effects when the group membership is unknown.
}
\keyword{datasets}
