% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HOMTEST.R
\name{HOMTEST}
\alias{HOMTEST}
\title{HOMTEST}
\usage{
HOMTEST(X, Y, Nfactors, Maxit = 100, tol = 0.001)
}
\arguments{
\item{X}{The (NT) times p design matrix, without an intercept where N=number of individuals, T=length of time series, p=number of explanatory variables.}

\item{Y}{The T times N panel of response where N=number of individuals, T=length of time series.}

\item{Nfactors}{A pre-specified number of common factors.}

\item{Maxit}{A maximum number of iterations in optimization. Default is 100.}

\item{tol}{Tolerance level of convergence. Default is 0.001.}
}
\value{
A list with the following components:
\itemize{
\item Coefficients: The estimated heterogeneous coefficients.
\item Factors: The estimated common factors across groups.
\item Loadings: The estimated factor loadings for the common factors.
\item pvalue: The p-value of the homogeneity test.
}
}
\description{
This function tests homogeneity of the regression coefficients in heterogeneous panel data models with interactive effects.
}
\examples{
N <- 100
P <- 200
R <- 2
p <- 2

LAM <- matrix(rnorm(P*R,0,1),nrow=P,ncol=R)
FAC <- matrix(rnorm(N*R,0,1),nrow=N,ncol=R)
AY <- FAC\%*\%t(LAM)+matrix(rnorm(N*P,0,1),nrow=N,ncol=P)
AX <- matrix(0,nrow=N*P,ncol=p)

for(j in 1:P){
  AX[(N*(j-1)+1):(N*j),1] <- 0.2+0.3*AY[,j]+matrix(rnorm(N,0,1),nrow=N)
  AX[(N*(j-1)+1):(N*j),2] <- 0.5+0.5*AY[,j]+matrix(rnorm(N,0,1),nrow=N)
}

B <- (1:p)/2-1
AB <- B\%*\%t(rep(1,len=P))

for(j in 1:P){AY[,j] <- AY[,j]+AX[(N*(j-1)+1):(N*j),]\%*\%AB[,j]}

fit <- HOMTEST(AX,AY,2)
}
\references{
Ando, T. and Bai, J. (2015) A simple new test for slope homogeneity in panel data models with interactive effects. Economics Letters, 136, 112-117.
}
